+++ Program.cs
@@ -43,6 +43,7 @@ builder.Services.AddScoped<IMonthlyFundingReviewService, MonthlyFundingReviewSer
 builder.Services.AddScoped<IReserveAccountSelectionService, ReserveAccountSelectionService>();
 builder.Services.AddScoped<IReservePotActionService, ReservePotActionService>();
 builder.Services.AddScoped<IReserveInterestForecastService, ReserveInterestForecastService>();
+builder.Services.AddScoped<IFinancialForecastService, FinancialForecastService>();
 builder.Services.AddScoped<FinanceRepository>();
 builder.Services.AddScoped<FinanceCalculator>();
 builder.Services.AddHttpClient<MarketPriceService>();